Initial commit: Password generator in Go
This commit marks the initial version of my
password generator written in Go.
The program generates random passwords of a
specified length using a set of characters.
It provides a secure and customizable way to
create strong passwords.

Features implemented in this commit:
- Generation of random passwords based on the
  specified length
- Error handling for any potential issues during
  password generation
- Copy the password to the clipboard

Future enhancements and additions will be made
in subsequent commits. This project aims to
provide a helpful tool for creating secure
passwords in Go. Excited to begin this journey!"